Add Dockerfile with node:slim-8
In order to test CD in gcloud with kubernetes it's needed to create a
docker image.

SQLite still not working in this image blue print. Probably in the
future it will be migrate to a cloud native solution since the SQL
database was a choice due the infrastructure limitations at time.